feat: Update product and tenant handling; fix typos and enhance search functionality

+++ webParentVer2/WebBackEndVer2/src/main/java/com/onlineStore/admin/MainController.java
@@ -144,7 +144,7 @@ public class MainController {
             model.addAttribute("pageTitle", "My Store Dashboard");
 
             if (tenantIdLong != null) {
-                Integer tenantId = tenantIdLong.intValue();
+                Long tenantId = tenantIdLong;
 
                 // Analytics - Top Viewed
                 java.util.List<Object[]> topViewed = analyticsService.getTopViewedProducts(tenantId, 5);
